@@ -2,14 +2,14 @@
   "name": "iconoir.com",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbo",
     "build": "next build",
     "start": "pnpm dlx serve out",
     "lint": "next lint"
   },
   "devDependencies": {
-    "@nodesecure/npm-registry-sdk": "^2.1.0",
-    "@octokit/rest": "^20.0.2",
+    "@nodesecure/npm-registry-sdk": "3.0.0",
+    "@octokit/rest": "21.0.2",
     "@react-aria/focus": "^3.15.0",
     "@react-aria/i18n": "^3.9.0",
     "@react-aria/slider": "^3.7.3",
@@ -19,27 +19,26 @@
     "@react-types/slider": "^3.7.0",
     "@types/animejs": "^3.1.12",
     "@types/lodash": "^4.14.202",
-    "@types/node": "^20.10.4",
+    "@types/node": "22.10.2",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.2.18",
     "@types/react-window": "^1.8.8",
-    "@types/remark-prism": "1.3.7",
     "animejs": "^3.2.2",
     "csvtojson": "^2.0.10",
     "eslint-config-next": "15.1.0",
     "iconoir-react": "workspace:*",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
-    "next": "^14.0.4",
-    "next-mdx-remote": "^4.4.1",
+    "next": "^15.1.0",
+    "next-mdx-remote": "^5.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-window": "^1.8.10",
-    "remark-gfm": "^3.0.1",
+    "remark-gfm": "4.0.0",
     "remark-prism": "^1.3.6",
     "scule": "^1.1.1",
     "styled-components": "^6.1.1",
-    "typescript": "^5.3.3",
+    "typescript": "5.7.2",
     "use-resize-observer": "^9.1.0"
   }
 }
